Discriminate between attributes in different namespaces
This commit aims to fix https://github.com/eminence/xmltree-rs/issues/13
(Attribute namespaces are not preserved). It changes the type of the
`attributes` map on the `Element` struct from a String -> String map to
an AttributeName -> String map. AttributeName is a re-export of the
xml::name::OwnedName type in xmltree-rs, which is a struct containing
the attribute's local name, namespace and prefix.

It is possible to search the `attributes` map using a fully initialised
AttributeName structure, but for convenience get_attribute(),
get_mut_attribute() and take_attribute() methods, modelled on the
*_element() equivalent are provided, which take an AttributePredicate.
Ready-made implementations taking either a local name or local name +
namespace are provided.